Restructure and added printing and saving
