+++ .gitignore
@@ -351,3 +351,4 @@ MigrationBackup/
 
 # Conforming to SRD
 DndUtils/CharacterGenerator/Data/Secrets/
+CharacterGenerator/Data/Secrets/
