+++ CharacterGenerator/Data/CharacterModel.cs
