+++ CharacterGenerator/Data/CharacterView.cs
