+++ CharacterGenerator/Data/IClass.cs
