+++ CharacterGenerator/Data/IFeat.cs
