+++ CharacterGenerator/Data/Race/Gnome.cs
