+++ CharacterGenerator/Data/Race/IRace.cs
