+++ CharacterGenerator/DiceRoller.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Numerics;
 using System.Text;
 
-namespace DndUtils
+namespace DndUtils.CharacterGenerator
 {
     static class DiceRoller
     {
