+++ CharacterGenerator/Pages/CharacterGeneratorASIPage.xaml
