+++ CharacterGenerator/Pages/CharacterGeneratorASIPage.xaml.cs
