+++ CharacterGenerator/Pages/CharacterGeneratorCustomCharacterPage.xaml
