+++ CharacterGenerator/Pages/CharacterGeneratorCustomCharacterPage.xaml.cs
