+++ CharacterGenerator/Pages/CharacterGeneratorHomePage.xaml
