+++ CharacterGenerator/Pages/CharacterGeneratorHomePage.xaml.cs
