+++ CharacterGenerator/Pages/CharacterGeneratorMain.xaml
