+++ CharacterGenerator/Pages/CharacterGeneratorMain.xaml.cs
