+++ DndUtils.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Blend for Visual Studio Version 16
+# Visual Studio Version 16
 VisualStudioVersion = 16.0.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{6876ADA7-4633-4277-9F82-588845F996DB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DndUtils", "DndUtils\DndUtils.csproj", "{B00DB65C-5835-447E-9F0B-6B2697E7658A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DndUtils", "DndUtils\DndUtils.csproj", "{B00DB65C-5835-447E-9F0B-6B2697E7658A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CharacterGenerator", "CharacterGenerator\CharacterGenerator.csproj", "{2613086B-A1DB-4E00-898B-A3EDD127E2C7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,6 +19,10 @@ Global
 		{B00DB65C-5835-447E-9F0B-6B2697E7658A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B00DB65C-5835-447E-9F0B-6B2697E7658A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B00DB65C-5835-447E-9F0B-6B2697E7658A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2613086B-A1DB-4E00-898B-A3EDD127E2C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2613086B-A1DB-4E00-898B-A3EDD127E2C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2613086B-A1DB-4E00-898B-A3EDD127E2C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2613086B-A1DB-4E00-898B-A3EDD127E2C7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
